// src/core/usuario/enum/permissoes_funcionalidades.enum.ts
export enum PERMISSOES_LOJA  {
    LOJA_VER_DASHBOARD = 'lojaVerDashboard',
    LOJA_VER_ATENDIMENTOS = 'lojaVerAtendimentos',
    LOJA_EDITAR_EXCLUIR_ATENDIMENTO = 'lojaEditarExcluirAtendimento',
    LOJA_VINCULAR_ATENDIMENTO_USUARIO = 'lojaVincularAtendimentoUsuario',
    LOJA_TRANSFERIR_ATENDIMENTO = 'lojaTransferirAtendimento',
    LOJA_VER_CHAT = 'lojaVerChat',
    LOJA_RESPONDER_CHAT = 'lojaResponderChat',
    LOJA_CADASTRAR_EDITAR_CLIENTES = 'lojaCadastrarEditarClientes',
    LOJA_PESQUISAR_CLIENTES = 'lojaPesquisarClientes',
    LOJA_GERENCIAR_USUARIOS = 'lojaGerenciarUsuarios',
    LOJA_GERENCIAR_CARGOS = 'lojaGerenciarCargos',
    LOJA_CONFIGURAR_INTEGRACOES = 'lojaConfigurarIntegracoes',
    LOJA_EDITAR_DADOS_DA_LOJA = 'lojaEditarDadosDaLoja',
    LOJA_GERENCIAR_SUSPENSOES = 'lojaGerenciarSuspensoes',
    LOJA_VER_TODOS_ATENDIMENTOS = 'lojaVerTodosAtendimentos',
    LOJA_CRIAR_ATENDIMENTO_MANUAL = 'lojaCriarAtendimentoManual',
    LOJA_REATRIBUIR_ATENDIMENTOS = 'lojaReatribuirAtendimentos',
    LOJA_GERENCIAR_EQUIPE = 'lojaGerenciarEquipe',
    LOJA_VER_ATENDIMENTOS_PROPRIOS = 'lojaVerAtendimentosProprios',
    LOJA_CRIAR_ATENDIMENTO_PROPRIO = 'lojaCriarAtendimentoProprio',
    LOJA_GERENCIAR_TAREFAS_PROPRIAS = 'lojaGerenciarTarefasProprias',
    LOJA_PRE_VENDEDOR_FINALIZAR_ATENDIMENTO = 'lojaPreVendedorFinalizarAtendimento',
    LOJA_VER_MENSAGENS_PADRAO = 'lojaVerMensagensPadrao',
    LOJA_CRIAR_MENSAGEM_PADRAO = 'lojaCriarMensagemPadrao',
    LOJA_EDITAR_MENSAGEM_PADRAO = 'lojaEditarMensagemPadrao',
    LOJA_DELETAR_MENSAGEM_PADRAO = 'lojaDeletarMensagemPadrao',
}

export enum PERMISSOES_AUTOPILOT  {
    AUTOPILOT_VER_DASHBOARD = 'autopilotVerDashboard',
    AUTOPILOT_RESPONDER_TICKETS = 'autopilotResponderTickets',
    AUTOPILOT_VER_TICKETS = 'autopilotVerTickets',
    AUTOPILOT_BLOQUEAR_ASSINANTE = 'autopilotBloquearAssinante',
    AUTOPILOT_VER_ASSINANTES = 'autopilotVerAssinantes',
    AUTOPILOT_EDITAR_ASSINANTES_ADICIONAR = 'autopilotEditarAssinantesAdicionar',
    AUTOPILOT_ALTERAR_PAINEL_REVENDA = 'autopilotAlterarPainelRevenda',
    AUTOPILOT_ATUALIZAR_PERMISSOES = 'autopilotAtualizarPermissoes',
    AUTOPILOT_CRIAR_USUARIO_ADMIN = 'autopilotCriarUsuarioAdmin',
    AUTOPILOT_VER_USUARIOS_ADMIN = 'autopilotVerUsuariosAdmin',
}

export const PERMISSOES_TODAS = {
    ...PERMISSOES_LOJA,
    ...PERMISSOES_AUTOPILOT,
}